// my-first-nextjs/src/app/types/ButtonData.tsx
// export interface ButtonData {
//   id: string
//   label: string
//   type: number
//   origin: 'noHeader' | 'type1' | 'type2'
//   current?: 'noHeader' | 'type1' | 'type2'
// }


// ./app/types/ButtonData.ts
export interface ButtonData {
  id: string
  label: string
  type: string
  origin: string
  current: string
}